Uncap damage to 64-bit, mob HP to long, and rework DoT/summon damage
Pairs with kaentake feature/damage-long; both must ship together (attack,
summon and DAMAGE_MONSTER packets changed shape).

- Mob HP is long end to end (Monster/MonsterStats/OverrideMonsterStats/
  ChangeableStats, MonsterListener, MapleMap.damageMonster). WZ maxHP is
  read as int or string; mob-hp-overrides.yaml + !reloadmobhp (gm3) set
  HP above the WZ int range without touching .wz/.img data.
- Attack lines are [int | crit bit][long] per line (C2S and S2C); the long
  is trusted only when the int is the INT_MAX sentinel. Lines, totals and
  DamageRank are long with saturating sums; int-only APIs clamp.
- Magnitude DAMAGE_HACK checks and the summon calcMaxDamage clamp are
  removed: they are int estimates that cannot model uncapped stats. "Too
  many lines" stays.
- Summons: [int][long] per target both ways.
- DAMAGE_MONSTER: [int][long][int attacker id]; every DoT tick is
  broadcast with its real value so the client can draw it with the
  attacker's damage skin. Poison/venom lose the 32,767 cap (status value
  on the wire stays a short).
- Ninja Ambush: tick = max base damage x damage% (64-bit), applied even
  when the opening hit is 0, and ticks may kill.
- Snipe is no longer rewritten server side (client computes it).
- Percent potions capped: Power Elixir family 99,999, others 49,999.
- !killall / !cake / !bosshp handle HP above 2.1B.

+++ src/main/java/client/command/commands/gm1/BossHpCommand.java
@@ -38,7 +38,7 @@ public class BossHpCommand extends Command {
         Character player = c.getPlayer();
         for (Monster monster : player.getMap().getAllMonsters()) {
             if (monster != null && monster.isBoss() && monster.getHp() > 0) {
-                long percent = monster.getHp() * 100L / monster.getMaxHp();
+                long percent = (long) (monster.getHp() * 100d / monster.getMaxHp()); // * 100L overflows past ~9.2e16 HP
                 String bar = "[";
                 for (int i = 0; i < 100; i++) {
                     bar += i < percent ? "|" : ".";
